+++ delta/sdk/src/main/scala/ch/epfl/bluebrain/nexus/delta/sdk/acls/model/AclState.scala
@@ -67,7 +67,7 @@ final case class AclState(
     ResourceF(
       id = uris.relativeAccessUri.toIri,
       uris = uris,
-      rev = rev.toLong,
+      rev = rev,
       types = types,
       deprecated = deprecated,
       createdAt = createdAt,
