+++ delta/sdk/src/main/scala/ch/epfl/bluebrain/nexus/delta/sdk/model/IdSegmentRef.scala
@@ -18,7 +18,7 @@ sealed trait IdSegmentRef extends Product with Serializable { self =>
   /**
     * Converts the current [[IdSegmentRef]] to an [[IdSegmentRef.Revision]]
     */
-  def toRev(rev: Long): IdSegmentRef.Revision = IdSegmentRef.Revision(self.value, rev)
+  def toRev(rev: Int): IdSegmentRef.Revision = IdSegmentRef.Revision(self.value, rev)
 
   /**
     * Converts the current [[IdSegmentRef]] to an [[IdSegmentRef.Latest]]
@@ -44,9 +44,9 @@ object IdSegmentRef {
   implicit def stringToIdSegmentRef(string: String): IdSegmentRef   = Latest(string)
 
   def apply(id: IdSegment): IdSegmentRef                               = Latest(id)
-  def apply(id: IdSegment, rev: Long): IdSegmentRef                    = Revision(id, rev)
+  def apply(id: IdSegment, rev: Int): IdSegmentRef                     = Revision(id, rev)
   def apply(id: IdSegment, tag: UserTag): IdSegmentRef                 = Tag(id, tag)
-  def fromRevOpt(id: IdSegment, revOpt: Option[Long]): IdSegmentRef    = revOpt.fold(apply(id))(Revision(id, _))
+  def fromRevOpt(id: IdSegment, revOpt: Option[Int]): IdSegmentRef     = revOpt.fold(apply(id))(Revision(id, _))
   def fromTagOpt(id: IdSegment, tagOpt: Option[UserTag]): IdSegmentRef = tagOpt.fold(apply(id))(Tag(id, _))
 
   /**
@@ -67,7 +67,7 @@ object IdSegmentRef {
   /**
     * A segment annotated with a rev.
     */
-  final case class Revision(value: IdSegment, rev: Long) extends IdSegmentRef
+  final case class Revision(value: IdSegment, rev: Int) extends IdSegmentRef
 
   /**
     * A segment annotated with a tag.
