+++ delta/sdk/src/main/scala/ch/epfl/bluebrain/nexus/delta/sdk/schemas/model/SchemaState.scala
@@ -78,7 +78,7 @@ final case class SchemaState(
     ResourceF(
       id = id,
       uris = ResourceUris.schema(project, id)(mappings, base),
-      rev = rev.toLong,
+      rev = rev,
       types = types,
       schema = schema,
       deprecated = deprecated,
